+++ CHANGELOG.md
@@ -20,6 +20,7 @@ https://mediaelch.github.io/mediaelch-doc/faq.html#where-are-mediaelchs-settings
  - Fix ADE scraper (#703, #725)
  - Fix TV shows always beeing reloaded (#732)
  - Fix incompatibilities with Kodi v17 NFO files (#719)
+ - Fix race-condition in DownloadManager (#766)
 
 ### Improvements